Fixed customfields checkbox default value handling
This patch tries to fix the following situation:

Currently it is not possible to set all of the checkboxes of a custom
fields to an unchecked state, if the default values for them are not
empty. Trying to do so, will cause the checkboxes to fallback to their
default value.