Added article tag: &lt;ion:article:is_online&gt;
This new article tag allows a condition on articles being online or not. 
I am not sure if there's a bug in the tag manager causing it to expand custom article types always (ignoring them being set "online" in the current language or not), or if the tag manager renders offline article's content only when they contain custom tags (as in my case) - anyway, this condition allows to check for articles being online and solve this problem.